Discovered that drcAssign mapped arguments "dist" and "cdist" to type "short", incorrectly.  Although distances don't usually exceed 16 bits, a minimum area DRC rule area value at a small feature size---like 0.18um using nanometer units for DRC rules--- can easily exceed this, and results definitely do not come out as expected. . .

+++ drc/DRCtech.c
@@ -872,7 +872,7 @@ DRCTechLine(sectionName, argc, argv)
 void
 drcAssign(cookie, dist, next, mask, corner, why, cdist, flags, planeto, planefrom)
     DRCCookie *cookie, *next;
-    short dist, cdist;
+    int dist, cdist;
     TileTypeBitMask *mask, *corner;
     char *why;
     int flags, planeto, planefrom;
